[rpc] Rework class hierarchy to prepare RPC server implementation
