+++ src/chargepoint/status/StatusManager.cpp
@@ -288,7 +288,7 @@ bool StatusManager::handleMessage(const ocpp::messages::ChangeAvailabilityReq& r
     }
     else
     {
-        error_code    = ocpp::rpc::IRpcClient::RPC_ERROR_PROPERTY_CONSTRAINT_VIOLATION;
+        error_code    = ocpp::rpc::IRpc::RPC_ERROR_PROPERTY_CONSTRAINT_VIOLATION;
         error_message = "Invalid connector id";
     }
     return ret;
