+++ src/messages/ExtendedTriggerMessage.cpp
@@ -17,7 +17,7 @@ along with OpenOCPP. If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include "ExtendedTriggerMessage.h"
-#include "IRpcClient.h"
+#include "IRpc.h"
 
 using namespace ocpp::types;
 
@@ -57,7 +57,7 @@ bool ExtendedTriggerMessageReqConverter::fromJson(const rapidjson::Value&    jso
     data.requestedMessage = MessageTriggerEnumTypeHelper.fromString(json["requestedMessage"].GetString());
     if (!ret)
     {
-        error_code = ocpp::rpc::IRpcClient::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
+        error_code = ocpp::rpc::IRpc::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
     }
     return ret;
 }
