+++ src/messages/GetCompositeSchedule.cpp
@@ -18,7 +18,7 @@ along with OpenOCPP. If not, see <http://www.gnu.org/licenses/>.
 
 #include "GetCompositeSchedule.h"
 #include "ChargingScheduleConverter.h"
-#include "IRpcClient.h"
+#include "IRpc.h"
 
 using namespace ocpp::types;
 
@@ -49,7 +49,7 @@ bool GetCompositeScheduleReqConverter::fromJson(const rapidjson::Value&  json,
     }
     if (!ret)
     {
-        error_code = ocpp::rpc::IRpcClient::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
+        error_code = ocpp::rpc::IRpc::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
     }
     return ret;
 }
@@ -82,7 +82,7 @@ bool GetCompositeScheduleConfConverter::fromJson(const rapidjson::Value&   json,
     }
     if (!ret && !error_code)
     {
-        error_code = ocpp::rpc::IRpcClient::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
+        error_code = ocpp::rpc::IRpc::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
     }
     return true;
 }
