+++ src/messages/GetDiagnostics.cpp
@@ -17,7 +17,7 @@ along with OpenOCPP. If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include "GetDiagnostics.h"
-#include "IRpcClient.h"
+#include "IRpc.h"
 #include "Url.h"
 
 using namespace ocpp::types;
@@ -43,7 +43,7 @@ bool GetDiagnosticsReqConverter::fromJson(const rapidjson::Value& json,
     ret = ret && extract(json, "stopTime", data.stopTime, error_message);
     if (!ret)
     {
-        error_code = ocpp::rpc::IRpcClient::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
+        error_code = ocpp::rpc::IRpc::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
     }
     return ret;
 }
