+++ src/messages/MeterValues.cpp
@@ -17,7 +17,7 @@ along with OpenOCPP. If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include "MeterValues.h"
-#include "IRpcClient.h"
+#include "IRpc.h"
 #include "MeterValueConverter.h"
 
 using namespace ocpp::types;
