+++ src/messages/RemoteStartTransaction.cpp
@@ -18,7 +18,7 @@ along with OpenOCPP. If not, see <http://www.gnu.org/licenses/>.
 
 #include "RemoteStartTransaction.h"
 #include "ChargingProfileConverter.h"
-#include "IRpcClient.h"
+#include "IRpc.h"
 
 using namespace ocpp::types;
 
@@ -54,7 +54,7 @@ bool RemoteStartTransactionReqConverter::fromJson(const rapidjson::Value&    jso
     }
     if (!ret)
     {
-        error_code = ocpp::rpc::IRpcClient::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
+        error_code = ocpp::rpc::IRpc::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
     }
     return ret;
 }
