+++ src/messages/TriggerMessage.cpp
@@ -17,7 +17,7 @@ along with OpenOCPP. If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include "TriggerMessage.h"
-#include "IRpcClient.h"
+#include "IRpc.h"
 
 using namespace ocpp::types;
 
@@ -55,7 +55,7 @@ bool TriggerMessageReqConverter::fromJson(const rapidjson::Value& json,
     data.requestedMessage = MessageTriggerHelper.fromString(json["requestedMessage"].GetString());
     if (!ret)
     {
-        error_code = ocpp::rpc::IRpcClient::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
+        error_code = ocpp::rpc::IRpc::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
     }
     return ret;
 }
