+++ src/messages/UpdateFirmware.cpp
@@ -17,7 +17,7 @@ along with OpenOCPP. If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include "UpdateFirmware.h"
-#include "IRpcClient.h"
+#include "IRpc.h"
 #include "Url.h"
 
 using namespace ocpp::types;
@@ -42,7 +42,7 @@ bool UpdateFirmwareReqConverter::fromJson(const rapidjson::Value& json,
     ret = ret && extract(json, "retrieveDate", data.retrieveDate, error_message);
     if (!ret)
     {
-        error_code = ocpp::rpc::IRpcClient::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
+        error_code = ocpp::rpc::IRpc::RPC_ERROR_TYPE_CONSTRAINT_VIOLATION;
     }
     return ret;
 }
