+++ src/messages/types/AuthorizationDataConverter.cpp
@@ -18,7 +18,7 @@ along with OpenOCPP. If not, see <http://www.gnu.org/licenses/>.
 
 #include "AuthorizationDataConverter.h"
 #include "Enums.h"
-#include "IRpcClient.h"
+#include "IRpc.h"
 #include "IdTagInfoConverter.h"
 
 using namespace ocpp::types;
