+++ src/rpc/CMakeLists.txt
@@ -5,6 +5,7 @@
 
 # Library target
 add_library(rpc STATIC
+    RpcBase.cpp
     RpcClient.cpp
 )
